zoom tutorials consolidation
