+++ public/index.html
@@ -54,19 +54,19 @@
     (<a href="https://leafletjs.com/examples/zoom-levels/" target="_blank">tutorial</a>)
     <ol type="a">
       <li>
-        <a href="zoom-zero.html">Zero</a>
+        <a href="tutorial/zoom/zero.html">Zero</a>
       </li>
       <li>
-        <a href="zoom-scale.html">Scale</a>
+        <a href="tutorial/zoom/scale.html">Scale</a>
       </li>
       <li>
-        <a href="zoom-discrete.html">Discrete</a>
+        <a href="tutorial/zoom/discrete.html">Discrete</a>
       </li>
       <li>
-        <a href="zoom-continuous.html">Continuous</a>
+        <a href="tutorial/zoom/continuous.html">Continuous</a>
       </li>
       <li>
-        <a href="zoom-delta.html">Delta</a>
+        <a href="tutorial/zoom/delta.html">Delta</a>
       </li>
     </ol>
   </li>
