+++ public/tutorial/zoom/script/discrete.js
@@ -5,11 +5,11 @@ import {
   createMap,
   getMapZoom,
   setMapZoom,
-} from './map-utils.js';
+} from '../../../script/map-utils.js';
 import {
   attributionCarto,
   urlTemplateCarto,
-} from './base-layers.js';
+} from '../../../script/base-layers.js';
 
 const map = createMap({
   center: [0, 0],
