+++ public/tutorial/zoom/script/scale.js
@@ -3,11 +3,11 @@ import {
   addTileLayer,
   createMap,
   setMapView,
-} from './map-utils.js';
+} from '../../../script/map-utils.js';
 import {
   attributionCarto,
   urlTemplateCarto,
-} from './base-layers.js';
+} from '../../../script/base-layers.js';
 
 const map = createMap({
   center: [0, 0],
