+++ public/tutorial/zoom/zero.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <title>Zoom levels: Zero</title>
+  <link href="../../style/theme.css" rel="stylesheet"/>
+  <link href="../../leaflet/leaflet.css" rel="stylesheet"/>
+  <link href="../../style/map/medium.css" rel="stylesheet"/>
+  <script src="script/zero.js" type="module"></script>
+</head>
+
+<body>
+<div id="map"></div>
+</body>
+
+</html>
