--- public/zoom-scale.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <title>Zoom levels: Scale</title>
-  <link href="style/theme.css" rel="stylesheet"/>
-  <link href="leaflet/leaflet.css" rel="stylesheet"/>
-  <link href="style/map/medium.css" rel="stylesheet"/>
-  <script src="script/zoom-scale.js" type="module"></script>
-</head>
-
-<body>
-<div id="map"></div>
-</body>
-
-</html>
